refactor(llm): move tool_search instruction into model_constructors
